// PP-CoE-CodeApp/src/generated/models/AadusersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export interface Aadusers {
  aaduserid: string;
  accountenabled?: boolean;
  accountenabledname?: string;
  businessphones?: string;
  city?: string;
  companyname?: string;
  createddatetime?: string;
  displayname?: string;
  givenname?: string;
  id: string;
  imaddresses?: string;
  jobtitle?: string;
  mail?: string;
  mobilephone?: string;
  officelocation?: string;
  postalcode?: string;
  preferredlanguage?: string;
  streetaddress?: string;
  surname?: string;
  userprincipalname?: string;
  usertype?: string;
}